refactor : detail.js 페이지 코드 리팩토링
detail.js페이지 코드 리팩토링을 통해
컴포넌트를  분할하여 detail.js파일에는 최소한의 컴포넌트만 남김

#ISSUES 8181
